+++ Rank2/ft_printf/ft_unsigned.c
@@ -6,13 +6,13 @@
 /*   By: sraiha <sraiha@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/09 13:58:11 by sraiha            #+#    #+#             */
-/*   Updated: 2024/12/11 15:19:09 by sraiha           ###   ########.fr       */
+/*   Updated: 2024/12/13 15:30:58 by sraiha           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_putunsigned(unsigned int nbr, size_t *counter)
+void	ft_putunsigned(unsigned int nbr, int *counter)
 {
 	char	*number;
